+++ lib/td_dd_web/views/grant_request_search_view.ex
@@ -10,6 +10,7 @@ defmodule TdDdWeb.GrantRequestSearchView do
 
     result
     |> with_filters(assigns)
+    |> with_scroll_id(assigns)
     |> add_permissions(permissions)
   end
 
@@ -17,11 +18,15 @@ defmodule TdDdWeb.GrantRequestSearchView do
     Map.put(resp, :_permissions, permissions)
   end
 
-  defp add_permissions(resp, _), do: resp
-
   defp with_filters(payload, %{filters: filters}) do
     Map.put(payload, :filters, filters)
   end
 
   defp with_filters(payload, _assigns), do: payload
+
+  defp with_scroll_id(payload, %{scroll_id: scroll_id}) do
+    Map.put(payload, :scroll_id, scroll_id)
+  end
+
+  defp with_scroll_id(payload, _assigns), do: payload
 end